Added redux and split the Register widget on slices(register slice don't work now)
