+++ src/entities/task/api/api.js
@@ -1,0 +1,5 @@
+import api from '../../../shared/api/base'
+
+export const taskAPI = {
+    getTask: (data) => api.get('/tasks', data)
+}
