+++ src/entities/user/api/api.js
@@ -1,0 +1,6 @@
+import api from '../../../shared/api/base'
+
+export const userAPI = {
+    login: (data) => api.post('/login', data),
+    registration: (data) => api.post('/registration', data),
+}
