+++ src/features/auth/login/model/loginSelectors.js
@@ -1,0 +1,2 @@
+export const selectLoginState = (state) => state.login;
+export const selectIsLoggedIn = (state) => selectLoginState(state).isLogin;
