+++ src/features/auth/login/ui/LoginForm.module.css
@@ -1,0 +1,114 @@
+.r_form {
+    display: flex;
+    flex-direction: column;
+
+    justify-content: center;
+    align-items: center;
+
+    width: 500px;
+    padding: 96px 0 96px 0;
+    background-color: rgb(229, 208, 172);
+    box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.4);
+    border-radius: 12px;
+
+    gap: 24px;
+}
+
+.inputs {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 8px;
+}
+
+.logo_label {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 16px;
+}
+
+.logo_img {
+    height: 96px;
+    width: auto;
+}
+
+span {
+    font-family: Arial, Helvetica, sans-serif;
+}
+
+.warningText {
+    opacity: 0;
+    height: 0;
+    overflow: hidden;
+    transition: opacity 0.3s ease, height 0.3s ease;
+    color: red;
+    font-size: x-small;
+    background-color: white;
+    padding: 0;
+    border: none;
+}
+
+.warningText.show {
+    opacity: 1;
+    height: 20px;
+    padding: 4px; /* Чтоб не схлопывался контент */
+    border-radius: 6px;
+    border-bottom: 2px solid red;
+}
+
+.warningText.changing {
+    animation: blinkBorder 0.5s ease-in-out 2, shake 0.3s ease-in-out;
+}
+
+@keyframes shake {
+    0%, 100% { transform: translateX(0); }
+    25% { transform: translateX(-3px); }
+    50% { transform: translateX(3px); }
+    75% { transform: translateX(-2px); }
+}
+
+@keyframes blinkBorder {
+
+    0%,
+    100% {
+        box-shadow: 0 0 0;
+    }
+
+    50% {
+        box-shadow: 0 2px 10px rgba(255, 0, 0, 0.5);
+    }
+}
+
+
+.label {
+    font-size: large;
+    font-weight: 700;
+    color: rgb(200, 0, 0);
+}
+
+.animated_field {
+    opacity: 0;
+    max-height: 0;
+    overflow: hidden;
+    transition: opacity 0.5s ease, max-height 0.5s ease;
+}
+
+.animated_field.show {
+    opacity: 1;
+    max-height: 100px;
+    /* Достаточно для появления поля */
+}
+
+.reg_butt {
+    color: rgb(200, 0, 0);
+    transition: 0.2s;
+    font-size: medium;
+    font-family: 'Gill Sans', 'Gill Sans MT', Calibri, 'Trebuchet MS', sans-serif;
+}
+
+.reg_butt:hover {
+    color: rgb(255, 0, 0);
+    transition: 0s;
+    cursor: pointer;
+}
