--- src/shared/api/APIs/authorisationAPI.js
@@ -1,5 +1,0 @@
-import api from '../base'
-
-export const authorisationAPI = {
-    login: (data) => api.post('/login', data)
-}
