--- src/shared/api/APIs/registrationAPI.js
@@ -1,5 +1,0 @@
-import api from '../base'
-
-export const registrationApi = {
-    registration: (data) => api.post('/registration', data)
-}
