--- src/shared/api/APIs/taskAPI.js
@@ -1,5 +1,0 @@
-import api from '../base'
-
-export const taskApi = {
-    getTask: (data) => api.get('/tasks', data)
-}
