+++ src/shared/index.js
@@ -1,6 +1,5 @@
 export { DefaultInput } from "./ui/DefaultInput/DefaultInput";
 export { DefaultButton } from "./ui/DefaultButton/DefaultButton";
-export { taskApi } from "./api/APIs/taskAPI.js"
 export { worker } from "./mocks/browser"
 export { validatePassword } from "./lib/passwordValidation.js"
